// frontend/src/api-generated/models/phieuKiemKeResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietKiemKeFlatResponse } from './chiTietKiemKeFlatResponse';

export interface PhieuKiemKeResponse {
  id?: number;
  idDonVi?: number;
  dotKiemKeId?: number;
  maDotKiemKe?: string;
  tenDotKiemKe?: string;
  maPhieuKiemKe?: string;
  idPhongBanKiemKe?: number;
  tenPhongBan?: string;
  tenNhanVienKiemKe?: string;
  trangThai?: string;
  thoiGianThucHien?: string;
  thoiGianTao?: string;
  danhSachChiTiet?: ChiTietKiemKeFlatResponse[];
}