// frontend/src/api-generated/models/executeKiemKeRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { UpdateLinhKienItem } from './updateLinhKienItem';
import type { UpdatePhanMemItem } from './updatePhanMemItem';
import type { UpdateThietBiItem } from './updateThietBiItem';

export interface ExecuteKiemKeRequest {
  isSubmit: boolean;
  danhSachThietBi?: UpdateThietBiItem[];
  danhSachLinhKien?: UpdateLinhKienItem[];
  danhSachPhanMem?: UpdatePhanMemItem[];
}